requirements.txt
cssselect
cx_Oracle
getgauge
dictdiffer
jsonpath_rw_ext
libaio
lxml
mergedeep
psycopg2-binary
pycountry
requests
smart-assertions
appdirs==1.4.4
baron==0.10.1
certifi==2022.9.14
charset-normalizer==2.1.1
cssselect==1.1.0
cx-Oracle==8.3.0
debugpy==1.6.3
decorator==5.1.1
dictdiffer==0.9.0
getgauge==0.3.17
grpcio==1.48.1
idna==3.4
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
libaio==0.9.1
lxml==4.9.1
mergedeep==1.3.4
pbr==5.10.0
ply==3.11
protobuf==4.21.6
psycopg2-binary==2.9.3
pycountry==22.3.5
redbaron==0.9.2
requests==2.28.1
rply==0.7.8
six==1.16.0
smart-assertions==1.0.2
urllib3==1.26.12